KeyguardIndication: fix fc for certain devices on lockscreen charging info
https://del.dog/unimenocuc.log

+++ packages/SystemUI/src/com/android/systemui/statusbar/KeyguardIndicationController.java
@@ -729,7 +729,7 @@ public class KeyguardIndicationController implements StateListener,
             }
             if (mChargingVoltage > 0) {
                 batteryInfo = (batteryInfo == "" ? "" : batteryInfo + " · ") +
-                        String.format("%.1f", (mChargingVoltage / 1000 / 1000)) + "V";
+                        String.format("%.1f", (float) (mChargingVoltage / 1000 / 1000)) + "V";
             }
             if (mTemperature > 0) {
                 String batteryTemp = PixeldustUtils.mccCheck(mContext) ?
